# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "description": "Public, serializable snapshot of a job.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_detail": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          },
          "id": {
            "pattern": "^job_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JobProgress"
            },
            "title": "Progress",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobProgress": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "fraction": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fraction"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "at",
          "stage",
          "message"
        ],
        "title": "JobProgress",
        "type": "object"
      },
...
      "JsonValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "MotionPlan": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/PlannedAction"
            },
            "maxItems": 24,
            "minItems": 1,
            "title": "Actions",
            "type": "array"
          },
          "contacts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContactDefinition"
            },
            "title": "Contacts",
            "type": "array"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "pattern": "^plan_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          },
          "scene_id": {
            "pattern": "^scene_[a-z0-9][a-z0-9_-]*$",
            "title": "Scene Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0.0",
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Schema Version",
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/MotionStyle",
            "default": {
              "confidence": 0.65,
              "energy": 0.35,
              "exaggeration": 0.2,
              "tempo": 1.0,
              "tension": 0.2
            }
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "sync": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SyncConstraint"
            },
            "title": "Sync",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlanWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "summary",
          "actions",
          "id",
          "scene_id"
        ],
        "title": "MotionPlan",
        "type": "object"
      },
      "MotionPlanApplyPatchRequest": {
        "description": "Deterministically apply a previously generated and previewed patch.",
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          },
          "patch": {
            "$ref": "#/components/schemas/MotionPlanPatch"
          }
        },
        "required": [
          "patch",
          "expected_revision"
        ],
        "title": "MotionPlanApplyPatchRequest",
        "type": "object"
      },
      "MotionPlanApplyPatchResult": {
        "properties": {
          "diff": {
            "items": {
              "type": "string"
            },
            "title": "Diff",
            "type": "array"
          },
          "document_revision": {
            "title": "Document Revision",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/MotionPlan"
          },
          "previous_plan": {
            "$ref": "#/components/schemas/MotionPlan"
          }
        },
        "required": [
          "document_revision",
          "plan",
          "previous_plan",
          "diff"
        ],
        "title": "MotionPlanApplyPatchResult",
        "type": "object"
      },
      "MotionPlanCompileRequest": {
        "properties": {
          "clip_name": {
            "default": "Planned motion",
            "minLength": 1,
            "title": "Clip Name",
            "type": "string"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "MotionPlanCompileRequest",
        "type": "object"
      },
      "MotionPlanGenerationRequest": {
        "description": "AI motion planning request (Ollama planner -> deterministic compiler).",
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          },
          "model": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "use_fixture": {
            "default": false,
            "title": "Use Fixture",
            "type": "boolean"
          }
        },
        "required": [
          "model",
          "expected_revision",
          "prompt"
        ],
        "title": "MotionPlanGenerationRequest",
        "type": "object"
      },
      "MotionPlanMutationResult": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/ProjectDocument"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "document",
          "revision"
        ],
        "title": "MotionPlanMutationResult",
        "type": "object"
      },
      "MotionPlanPatch": {
        "additionalProperties": false,
        "description": "A correction against stable action ids, never a new unrelated plan.",
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/PatchOperation"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlanWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "summary",
          "operations"
        ],
        "title": "MotionPlanPatch",
        "type": "object"
      },
      "MotionPlanPatchRequest": {
        "description": "Ask the model for a scoped correction patch (plan.md \u00a78.6).",
        "properties": {
          "action_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Action Ids",
            "type": "array"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          },
          "instruction": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          },
          "model": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "time_range": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Range"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "model",
          "expected_revision",
          "instruction"
        ],
        "title": "MotionPlanPatchRequest",
        "type": "object"
      },
      "MotionPlanRead": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/MotionPlan"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "plan",
          "revision"
        ],
        "title": "MotionPlanRead",
        "type": "object"
      },
      "MotionPlanValidationRead": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Errors",
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/PlanWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "errors",
          "warnings"
        ],
        "title": "MotionPlanValidationRead",
        "type": "object"
      },
      "MotionPlanWrite": {
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/MotionPlan"
          }
        },
        "required": [
          "plan",
          "expected_revision"
        ],
        "title": "MotionPlanWrite",
        "type": "object"
      },
...
      "ProjectDocument": {
        "additionalProperties": false,
        "properties": {
          "asset_manifest": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AssetManifestEntry"
            },
            "title": "Asset Manifest",
            "type": "array"
          },
          "characters": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CharacterDefinition"
            },
            "title": "Characters",
            "type": "array"
          },
          "clips": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AnimationClip"
            },
            "title": "Clips",
            "type": "array"
          },
          "engine_version": {
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Engine Version",
            "type": "string"
          },
          "format": {
            "pattern": "^rigstory-project$",
            "title": "Format",
            "type": "string"
          },
          "generation_records": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GenerationRecord"
            },
            "title": "Generation Records",
            "type": "array"
          },
          "motion_plans": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MotionPlan"
            },
            "title": "Motion Plans",
            "type": "array"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          },
          "scenes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SceneDefinition"
            },
            "title": "Scenes",
            "type": "array"
          },
          "schema_version": {
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "format",
          "schema_version",
          "engine_version",
          "project"
        ],
        "title": "ProjectDocument",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RigStory Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/motion-plans/{plan_id}": {
      "get": {
        "operationId": "get_motion_plan_api_v1_motion_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionPlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Motion Plan",
        "tags": [
          "motion-plans"
        ]
      },
      "patch": {
        "description": "Persist a user edit of the plan (parameters, ordering, style).",
        "operationId": "update_motion_plan_api_v1_motion_plans__plan_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotionPlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionPlanMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Motion Plan",
        "tags": [
          "motion-plans"
        ]
      }
    },
    "/api/v1/motion-plans/{plan_id}/apply-patch": {
      "post": {
        "description": "Deterministically apply a previewed patch. Undo by re-saving the\nreturned ``previous_plan`` through PATCH /motion-plans/{plan_id}.",
        "operationId": "apply_motion_plan_patch_api_v1_motion_plans__plan_id__apply_patch_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotionPlanApplyPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionPlanApplyPatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Motion Plan Patch",
        "tags": [
          "motion-plans"
        ]
      }
    },
    "/api/v1/motion-plans/{plan_id}/compile": {
      "post": {
        "description": "Start a deterministic compile job for a stored plan (202 + job).",
        "operationId": "compile_motion_plan_route_api_v1_motion_plans__plan_id__compile_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotionPlanCompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Motion Plan Route",
        "tags": [
          "motion-plans"
        ]
      }
    },
    "/api/v1/motion-plans/{plan_id}/patch": {
      "post": {
        "description": "Ask the model for a scoped correction patch; returns a preview job.",
        "operationId": "patch_motion_plan_api_v1_motion_plans__plan_id__patch_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotionPlanPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Motion Plan",
        "tags": [
          "motion-plans"
        ]
      }
    },
    "/api/v1/motion-plans/{plan_id}/validate": {
      "post": {
        "operationId": "validate_motion_plan_route_api_v1_motion_plans__plan_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionPlanValidationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Motion Plan Route",
        "tags": [
          "motion-plans"
        ]
      }
    },
...
    "/api/v1/scenes/{scene_id}/motion-plans/generate": {
      "post": {
        "description": "Start an AI motion-planning job (specs \u00a723.5, returns 202 + a job).",
        "operationId": "generate_motion_plan_api_v1_scenes__scene_id__motion_plans_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotionPlanGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Motion Plan",
        "tags": [
          "motion-plans"
        ]
      }
    },
...
  }
}